+++ src/test/java/org/zalando/fahrschein/InMemoryCursorManagerTest.java
@@ -1,4 +1,8 @@
-package org.zalando.fahrschein.cursormanager;
+package org.zalando.fahrschein;
+
+import org.zalando.fahrschein.CursorManager;
+import org.zalando.fahrschein.InMemoryCursorManager;
+import org.zalando.fahrschein.jdbc.AbstractCursorManagerTest;
 
 public class InMemoryCursorManagerTest extends AbstractCursorManagerTest {
     private CursorManager cursorManager = new InMemoryCursorManager();
