+++ src/test/java/org/zalando/fahrschein/jdbc/PersistentCursorManagerIT.java
@@ -1,4 +1,4 @@
-package org.zalando.fahrschein.cursormanager;
+package org.zalando.fahrschein.jdbc;
 
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,6 +6,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.transaction.TransactionConfiguration;
 import org.springframework.transaction.annotation.Transactional;
+import org.zalando.fahrschein.CursorManager;
 import org.zalando.fahrschein.LocalPostgresConfiguration;
 
 @RunWith(SpringJUnit4ClassRunner.class)
